+++ App/Modules/Blog/Views/front/category.php
@@ -4,7 +4,7 @@ $blogBase    = $rawBlogBase === '' ? '' : '/' . $rawBlogBase;
 ?>
 <div class="container blog-category">
   <nav class="cat-breadcrumb">
-    <a href="<?php echo $baseUrl . ($blogBase !== '' ? $blogBase : '/'); ?>"><?php echo htmlspecialchars($settings['blog_title'] ?? 'Blog'); ?></a>
+    <a href="<?php echo htmlspecialchars(site_path($baseUrl, $blogBase !== '' ? $blogBase : '/')); ?>"><?php echo htmlspecialchars($settings['blog_title'] ?? 'Blog'); ?></a>
     <span class="sep">/</span>
     <span><?php echo htmlspecialchars($category['name']); ?></span>
   </nav>
@@ -25,7 +25,7 @@ $blogBase    = $rawBlogBase === '' ? '' : '/' . $rawBlogBase;
     <?php foreach ($posts as $post): ?>
     <li class="post-item">
       <h2 class="post-title">
-        <a href="<?php echo $baseUrl . $blogBase; ?>/<?php echo htmlspecialchars($post['slug']); ?>">
+        <a href="<?php echo htmlspecialchars(site_path($baseUrl, $blogBase . '/' . $post['slug'])); ?>">
           <?php echo htmlspecialchars($post['title']); ?>
         </a>
       </h2>
@@ -40,7 +40,7 @@ $blogBase    = $rawBlogBase === '' ? '' : '/' . $rawBlogBase;
       <?php if (!empty($post['excerpt'])): ?>
       <p class="post-excerpt"><?php echo htmlspecialchars($post['excerpt']); ?></p>
       <?php endif; ?>
-      <a href="<?php echo $baseUrl . $blogBase; ?>/<?php echo htmlspecialchars($post['slug']); ?>" class="read-more">Read more &rarr;</a>
+      <a href="<?php echo htmlspecialchars(site_path($baseUrl, $blogBase . '/' . $post['slug'])); ?>" class="read-more">Read more &rarr;</a>
     </li>
     <?php endforeach; ?>
   </ul>